+++ tests/milvus_python_test/test_search_vectors.py
@@ -16,8 +16,6 @@ dim = 128
 table_id = "test_search"
 add_interval_time = 2
 vectors = gen_vectors(6000, dim)
-# vectors /= numpy.linalg.norm(vectors)
-# vectors = vectors.tolist()
 nprobe = 1
 epsilon = 0.001
 tag = "1970-01-01"
@@ -25,7 +23,7 @@ raw_vectors, binary_vectors = gen_binary_vectors(6000, dim)
 
 
 class TestSearchBase:
-    def init_data(self, connect, table, nb=6000):
+    def init_data(self, connect, table, nb=6000, partition_tags=None):
         '''
         Generate vectors and add it in table, before search vectors
         '''
@@ -34,13 +32,16 @@ class TestSearchBase:
             add_vectors = vectors
         else:  
             add_vectors = gen_vectors(nb, dim)
-            # add_vectors /= numpy.linalg.norm(add_vectors)
-            # add_vectors = add_vectors.tolist()
-        status, ids = connect.add_vectors(table, add_vectors)
+        if partition_tags is None:
+            status, ids = connect.add_vectors(table, add_vectors)
+            assert status.OK()
+        else:
+            status, ids = connect.add_vectors(table, add_vectors, partition_tags)
+            assert status.OK()
         sleep(add_interval_time)
         return add_vectors, ids
 
-    def init_binary_data(self, connect, table, nb=6000, insert=True):
+    def init_binary_data(self, connect, table, nb=6000, insert=True, partition_tags=None):
         '''
         Generate vectors and add it in table, before search vectors
         '''
@@ -52,10 +53,13 @@ class TestSearchBase:
             add_raw_vectors = raw_vectors
         else:  
             add_raw_vectors, add_vectors = gen_binary_vectors(nb, dim)
-            # add_vectors /= numpy.linalg.norm(add_vectors)
-            # add_vectors = add_vectors.tolist()
         if insert is True:
-            status, ids = connect.add_vectors(table, add_vectors)
+            if partition_tags is None:
+                status, ids = connect.add_vectors(table, add_vectors)
+                assert status.OK()
+            else:
+                status, ids = connect.add_vectors(table, add_vectors, partition_tags)
+                assert status.OK()
             sleep(add_interval_time)
         return add_raw_vectors, add_vectors, ids
 
@@ -169,8 +173,7 @@ class TestSearchBase:
         '''
         index_params = get_simple_index_params
         logging.getLogger().info(index_params)
-        partition_name = gen_unique_str()
-        status = connect.create_partition(table, partition_name, tag)
+        status = connect.create_partition(table, tag)
         vectors, ids = self.init_data(connect, table)
         status = connect.create_index(table, index_params)
         query_vec = [vectors[0]]
@@ -195,14 +198,13 @@ class TestSearchBase:
         '''
         index_params = get_simple_index_params
         logging.getLogger().info(index_params)
-        partition_name = gen_unique_str()
-        status = connect.create_partition(table, partition_name, tag)
+        status = connect.create_partition(table, tag)
         vectors, ids = self.init_data(connect, table)
         status = connect.create_index(table, index_params)
         query_vec = [vectors[0]]
         top_k = 10
         nprobe = 1
-        status, result = connect.search_vectors(partition_name, top_k, nprobe, query_vec, partition_tags=[tag])
+        status, result = connect.search_vectors(table, top_k, nprobe, query_vec, partition_tags=[tag])
         logging.getLogger().info(result)
         assert status.OK()
         assert len(result) == 0
@@ -215,9 +217,8 @@ class TestSearchBase:
         '''
         index_params = get_simple_index_params
         logging.getLogger().info(index_params)
-        partition_name = gen_unique_str()
-        status = connect.create_partition(table, partition_name, tag)
-        vectors, ids = self.init_data(connect, partition_name)
+        status = connect.create_partition(table, tag)
+        vectors, ids = self.init_data(connect, table, partition_tags=tag)
         status = connect.create_index(table, index_params)
         query_vec = [vectors[0]]
         top_k = 10
@@ -234,10 +235,6 @@ class TestSearchBase:
         assert len(result[0]) == min(len(vectors), top_k)
         assert check_result(result[0], ids[0])
         assert result[0][0].distance <= epsilon
-        status, result = connect.search_vectors(partition_name, top_k, nprobe, query_vec, partition_tags=[tag])
-        logging.getLogger().info(result)
-        assert status.OK()
-        assert len(result) == 0
 
     def test_search_l2_index_params_partition_C(self, connect, table, get_simple_index_params):
         '''
@@ -247,9 +244,8 @@ class TestSearchBase:
         '''
         index_params = get_simple_index_params
         logging.getLogger().info(index_params)
-        partition_name = gen_unique_str()
-        status = connect.create_partition(table, partition_name, tag)
-        vectors, ids = self.init_data(connect, partition_name)
+        status = connect.create_partition(table, tag)
+        vectors, ids = self.init_data(connect, table, partition_tags=tag)
         status = connect.create_index(table, index_params)
         query_vec = [vectors[0]]
         top_k = 10
@@ -269,9 +265,8 @@ class TestSearchBase:
         '''
         index_params = get_simple_index_params
         logging.getLogger().info(index_params)
-        partition_name = gen_unique_str()
-        status = connect.create_partition(table, partition_name, tag)
-        vectors, ids = self.init_data(connect, partition_name)
+        status = connect.create_partition(table, tag)
+        vectors, ids = self.init_data(connect, table, partition_tags=tag)
         status = connect.create_index(table, index_params)
         query_vec = [vectors[0]]
         top_k = 10
@@ -290,12 +285,10 @@ class TestSearchBase:
         new_tag = "new_tag"
         index_params = get_simple_index_params
         logging.getLogger().info(index_params)
-        partition_name = gen_unique_str()
-        new_partition_name = gen_unique_str()
-        status = connect.create_partition(table, partition_name, tag)
-        status = connect.create_partition(table, new_partition_name, new_tag)
-        vectors, ids = self.init_data(connect, partition_name)
-        new_vectors, new_ids = self.init_data(connect, new_partition_name, nb=6001)
+        status = connect.create_partition(table, tag)
+        status = connect.create_partition(table, new_tag)
+        vectors, ids = self.init_data(connect, table, partition_tags=tag)
+        new_vectors, new_ids = self.init_data(connect, table, nb=6001, partition_tags=new_tag)
         status = connect.create_index(table, index_params)
         query_vec = [vectors[0], new_vectors[0]]
         top_k = 10
@@ -325,12 +318,10 @@ class TestSearchBase:
         new_tag = "new_tag"
         index_params = get_simple_index_params
         logging.getLogger().info(index_params)
-        partition_name = gen_unique_str()
-        new_partition_name = gen_unique_str()
-        status = connect.create_partition(table, partition_name, tag)
-        status = connect.create_partition(table, new_partition_name, new_tag)
-        vectors, ids = self.init_data(connect, partition_name)
-        new_vectors, new_ids = self.init_data(connect, new_partition_name, nb=6001)
+        status = connect.create_partition(table, tag)
+        status = connect.create_partition(table, new_tag)
+        vectors, ids = self.init_data(connect, table, partition_tags=tag)
+        new_vectors, new_ids = self.init_data(connect, table, nb=6001, partition_tags=new_tag)
         status = connect.create_index(table, index_params)
         query_vec = [vectors[0], new_vectors[0]]
         top_k = 10
@@ -378,8 +369,7 @@ class TestSearchBase:
         '''
         index_params = get_simple_index_params
         logging.getLogger().info(index_params)
-        partition_name = gen_unique_str()
-        status = connect.create_partition(ip_table, partition_name, tag)
+        status = connect.create_partition(ip_table, tag)
         vectors, ids = self.init_data(connect, ip_table)
         status = connect.create_index(ip_table, index_params)
         query_vec = [vectors[0]]
@@ -404,9 +394,8 @@ class TestSearchBase:
         '''
         index_params = get_simple_index_params
         logging.getLogger().info(index_params)
-        partition_name = gen_unique_str()
-        status = connect.create_partition(ip_table, partition_name, tag)
-        vectors, ids = self.init_data(connect, partition_name)
+        status = connect.create_partition(ip_table, tag)
+        vectors, ids = self.init_data(connect, ip_table, partition_tags=tag)
         status = connect.create_index(ip_table, index_params)
         query_vec = [vectors[0]]
         top_k = 10
@@ -417,11 +406,6 @@ class TestSearchBase:
         assert len(result[0]) == min(len(vectors), top_k)
         assert check_result(result[0], ids[0])
         assert abs(result[0][0].distance - numpy.inner(numpy.array(query_vec[0]), numpy.array(query_vec[0]))) <= gen_inaccuracy(result[0][0].distance)
-        status, result = connect.search_vectors(partition_name, top_k, nprobe, query_vec)
-        logging.getLogger().info(result)
-        assert status.OK()
-        assert len(result[0]) == min(len(vectors), top_k)
-        assert check_result(result[0], ids[0])
 
     @pytest.mark.level(2)
     def test_search_vectors_without_connect(self, dis_connect, table):
@@ -899,7 +883,7 @@ class TestSearchParamsInvalid(object):
         nprobe = 1 
         query_vecs = gen_vectors(1, dim)
         with pytest.raises(Exception) as e:
-            status, result = connect.search_vectors(table_name, top_k, nprobe, query_vecs, partition_tags="tag")
+            status, result = connect.search_vectors(table, top_k, nprobe, query_vecs, partition_tags="tag")
 
     """
     Test search table with invalid top-k
