+++ tests/milvus_python_test/test_table_count.py
@@ -65,16 +65,15 @@ class TestTableCount:
         expected: the count is equal to the length of vectors
         '''
         nb = add_vectors_nb
-        partition_name = gen_unique_str()
         vectors = gen_vectors(nb, dim)
-        status = connect.create_partition(table, partition_name, tag)
+        status = connect.create_partition(table, tag)
         assert status.OK()
         res = connect.add_vectors(table_name=table, records=vectors, partition_tag=tag)
         time.sleep(add_time_interval)
         status, res = connect.get_table_row_count(table)
         assert res == nb
 
-    def test_table_rows_count_multi_partitions_A(self, connect, table, add_vectors_nb):
+    def test_table_rows_count_multi_partitions(self, connect, table, add_vectors_nb):
         '''
         target: test table rows_count is correct or not
         method: create table, create partitions and add vectors in it,
@@ -83,28 +82,25 @@ class TestTableCount:
         '''
         new_tag = "new_tag"
         nb = add_vectors_nb
-        partition_name = gen_unique_str()
-        new_partition_name = gen_unique_str()
         vectors = gen_vectors(nb, dim)
-        status = connect.create_partition(table, partition_name, tag)
-        status = connect.create_partition(table, new_partition_name, new_tag)
+        status = connect.create_partition(table, tag)
+        status = connect.create_partition(table, new_tag)
         assert status.OK()
         res = connect.add_vectors(table_name=table, records=vectors)
         time.sleep(add_time_interval)
         status, res = connect.get_table_row_count(table)
         assert res == nb
 
+    '''
     def test_table_rows_count_multi_partitions_B(self, connect, table, add_vectors_nb):
-        '''
+        
         target: test table rows_count is correct or not
         method: create table, create partitions and add vectors in one of the partitions,
             assert the value returned by get_table_row_count method is equal to length of vectors
         expected: the count is equal to the length of vectors
-        '''
+        
         new_tag = "new_tag"
         nb = add_vectors_nb
-        partition_name = gen_unique_str()
-        new_partition_name = gen_unique_str()
         vectors = gen_vectors(nb, dim)
         status = connect.create_partition(table, partition_name, tag)
         status = connect.create_partition(table, new_partition_name, new_tag)
@@ -117,12 +113,12 @@ class TestTableCount:
         assert res == 0
 
     def test_table_rows_count_multi_partitions_C(self, connect, table, add_vectors_nb):
-        '''
+        
         target: test table rows_count is correct or not
         method: create table, create partitions and add vectors in one of the partitions,
             assert the value returned by get_table_row_count method is equal to length of vectors
         expected: the table count is equal to the length of vectors
-        '''
+        
         new_tag = "new_tag"
         nb = add_vectors_nb
         partition_name = gen_unique_str()
@@ -140,6 +136,8 @@ class TestTableCount:
         assert res == nb
         status, res = connect.get_table_row_count(table)
         assert res == nb * 2
+
+    '''
 
     def test_table_rows_count_after_index_created(self, connect, table, get_simple_index_params):
         '''
